// src/app/models/interfaces.ts
export interface LoginInfo {
  id: string,
  user: string,
  email: string,
  token: string,
  expire: number,
  partyId: string,
  username: string,
  roles: any[],
  organizations: any[],
  logged_as: string
}